requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
brotlipy=0.7.0=py39h63b48b0_1004
bzip2=1.0.8=h0d85af4_4
ca-certificates=2022.9.24=h033912b_0
certifi=2022.9.24=pyhd8ed1ab_0
cffi=1.15.1=py39hae9ecf2_0
charset-normalizer=2.1.1=pyhd8ed1ab_0
cryptography=37.0.4=py39h9c2a9ce_0
freetype=2.12.1=h3f81eb7_0
idna=3.4=pyhd8ed1ab_0
joblib=1.2.0=pyhd8ed1ab_0
jpeg=9e=hac89ed1_2
lcms2=2.12=h577c468_0
lerc=4.0.0=hb486fe8_0
libblas=3.9.0=16_osx64_openblas
libcblas=3.9.0=16_osx64_openblas
libcxx=14.0.6=hccf4f1f_0
libdeflate=1.14=hb7f2c08_0
libffi=3.4.2=h0d85af4_5
libgfortran=5.0.0=9_5_0_h97931a8_26
libgfortran5=11.3.0=h082f757_26
liblapack=3.9.0=16_osx64_openblas
libopenblas=0.3.21=openmp_h429af6e_3
libpng=1.6.38=ha978bb4_0
libsqlite=3.39.3=ha978bb4_0
libtiff=4.4.0=hdb44e8a_4
libwebp-base=1.2.4=h775f41a_0
libxcb=1.13=h0d85af4_1004
libzlib=1.2.12=hfd90126_3
llvm-openmp=15.0.4=h61d9ccf_0
ncurses=6.3=h96cf925_1
numpy=1.23.4=py39hdfa1d0c_1
openjpeg=2.5.0=h5d0d7b0_1
openssl=1.1.1s=hfd90126_0
pandas=1.5.1=py39hecff1ad_1
pillow=9.2.0=py39h4d560c1_2
pip=22.2.2=pyhd8ed1ab_0
pthread-stubs=0.4=hc929b4f_1001
py-getch=1.0.1=pypi_0
pyautogui=0.9.48=py39h6e9494a_2
pycparser=2.21=pyhd8ed1ab_0
pymsgbox=1.0.9=pyhd8ed1ab_1
pyobjc-core=8.5.1=py39h0504f80_0
pyobjc-framework-cocoa=8.5=py39hf3ca4cb_0
pyobjc-framework-quartz=8.5.1=py39h6e9494a_0
pyopenssl=22.0.0=pyhd8ed1ab_1
pyscreeze=0.1.28=pyhd8ed1ab_1
pysocks=1.7.1=pyha2e5f31_6
python=3.9.13=h57e37ff_0_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.9=2_cp39
pytweening=1.0.4=pyhd8ed1ab_0
pytz=2022.6=pyhd8ed1ab_0
readline=8.1.2=h3899abd_0
requests=2.28.1=pyhd8ed1ab_1
scikit-learn=1.1.3=py39hdbdcc14_1
scipy=1.9.3=py39h8a15683_2
setuptools=65.4.0=pyhd8ed1ab_0
six=1.16.0=pyh6c4a22f_0
sqlite=3.39.3=h9ae0607_0
threadpoolctl=3.1.0=pyh8a188c0_0
tk=8.6.12=h5dbffcc_0
tzdata=2022d=h191b570_0
urllib3=1.26.11=pyhd8ed1ab_0
wheel=0.37.1=pyhd8ed1ab_0
xorg-libxau=1.0.9=h35c211d_0
xorg-libxdmcp=1.1.3=h35c211d_0
xz=5.2.6=h775f41a_0
zstd=1.5.2=hfa58983_4
